Initial implementation of a Chrome extension
This extension allows for arbitrary web pages to be turned into
Dallinger experiments.

It requests all accesses to https://*.herokuapp.com/* and browser
history. It uses that to listen to when the user opens a new tab,
and if it matches https://*.herokuapp.com/* it checks if the URL
starts with https://dlgr-. If it does, it is taken to be a Dallinger
experiment and an onclick handler is added for a button to allow
the experiment to request access to runtime specified websites.
The user would then see a permission request dialogue.

Once the user accepts that permission request, if tab navigates
to a matching URL the inject-mitm.js script will be added as a
content script. This will take care of inserting tracking javascript
once it is created. It also activates the page action which displays
the Dallinger logo in colour in the toolbar. If the user clicks the
logo when it is shaded they will see a list of current listeners
and be presented with the option to end the listening.